% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Klargjør trigger databasen}
\usage{
init(admin_email = "admin@example.com")
}
\arguments{
\item{admin_email}{epostadresse som varsel sendes til dersom programvaren
ikke virker som den skal.}
}
\value{
stien der databasen og programvaren finnes.
}
\description{
Klargjør triggerdatabasen. Tester også enkle transaksjoner i databasen.
}
\examples{

locate_database("/si/testing/trigger/")
init()

}
