% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_database.R
\name{locate_database}
\alias{locate_database}
\title{Assign the location of the trigger-database.}
\usage{
locate_database(path = "/si/computed/trigger/")
}
\arguments{
\item{path}{\code{path} where the database is located.}
}
\value{
intput is returned invisible
}
\description{
Assign the location of the trigger-database. This is the databse where jobs
and runs are registred. The function places this information in options.
}
