% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql_file.R
\name{read_sql_file}
\alias{read_sql_file}
\title{Les SQL}
\usage{
read_sql_file(filsti, params = NULL)
}
\arguments{
\item{filsti}{character string. Må angi en fil med endelse .sql eller .SQL.}

\item{params}{list. Optional. A list with the parameters in the corresponding sql file.}
}
\value{
character string. Returnrerer SQL i filstien
}
\description{
Leser innholdet fra en SQL-fil. Filen må være en .sql fil.
}
\keyword{internal}
