% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_job.R
\name{register_job}
\alias{register_job}
\title{Register job with trigger}
\usage{
register_job(name, frequency, job_description = " ", ...)
}
\arguments{
\item{name}{\code{character}. Name of the job. All characters are allowed.}

\item{frequency}{\code{daily()} or \code{monthly()}. Assigned from \code{run_frequencies}.}
}
\value{
invisibly the record in jobs
}
\description{
Register job wit trigger. This ensures that the job has uniue run-ids, and
that runs can be associated. It is a pre-requeisit for running a job with
\code{trigger}.
}
\examples{

x <- register_job("gralen_snm",run_frequencies$monthly())

}
