% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_frequency.R
\docType{data}
\name{run_frequencies}
\alias{run_frequencies}
\title{Run frequencies}
\format{
An object of class \code{list} of length 2.
}
\usage{
run_frequencies
}
\description{
List of run frequencies. Used when regisering a job.
}
\examples{
run_frequencies$daily()
run_frequencies$monthly()

}
\keyword{datasets}
