% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_job_runner.R
\name{set_up_job_runner}
\alias{set_up_job_runner}
\title{Set up job-runner}
\usage{
set_up_job_runner(name, alerts)
}
\arguments{
\item{name}{}

\item{alerts}{}
}
\value{
\code{function} which can be used to
}
\description{
Set up a job runner.
}
\examples{

init()
register_job("test",run_frequencies$daily())
job_runner <- set_up_job_runner(name = "test",alerts = c())
job_runner({
  Sys.sleep(1)
})
job_runner <- set_up_job_runner(name = "test",alerts = c())
job_runner({
  Sys.sleep(1)
  warning("Warning event")
})
job_runner <- set_up_job_runner(name = "test",alerts = c())
job_runner({
  Sys.sleep(2)
  stop("Error event")
})
get_table("runs",get_trigger_dbc())
get_table("events",get_trigger_dbc())
}
